collision changes,highscore,rotation lock,death
ive added a highscore system that tracks your current highscore, i also changed the way the score is done and made the whole thing a little easier...( it will still go banana after like 20 lines cleared.

also changed the y collision a little , so that you can move around when you are about to hit a unit or the ground , but when the next "frame" of moving down is done it will stop. this makes moving pieces into tight gaps easier / possible. 

Also rotation is now locked when a collision on the x or y axis is detected. Unfortunately it also locks the rotation even if there is no harm to rotating .
Made the 4*1 to a 3*1 so that it doesn't give me any trouble with my poor rotation collision detection.

AAAnd i also added a loose condition